refactor: refactoring gatewayConnection
