+++ DGCAVerifier/SupportingFiles/Resources/Configurations/Production.xcconfig
@@ -2,6 +2,3 @@
 baseUrl = https:\/\/get.dgc.gov.it/v1/dgc/
 certificateEvaluator = get.dgc.gov.it
 certificateFilename = get-dgc-gov-it.der
-updateEndpoint = signercertificate/update
-statusEndpoint = signercertificate/status
-settingsEndpoint = settings
